Update README.md
The password manager built in Python offers a secure and convenient solution for storing and managing passwords. By utilizing the Fernet encryption algorithm and employing proper security measures, it ensures that user passwords remain protected even in the event of unauthorized access. The password manager provides an easy-to-use interface for adding and viewing passwords, enhancing the overall user experience while prioritizing security and privacy.